% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grimon.R, R/zzz.R
\docType{package}
\name{grimon}
\alias{grimon}
\alias{grimon}
\alias{grimon-package}
\title{grimon}
\usage{
grimon(x, format = "wide", segment_mat = NULL, col = "black",
  label = NULL, optimize_coordinates = FALSE, maxiter = 1000, initT = 1,
  alpha = NULL, progress = FALSE, norm = TRUE, norm_scale = NULL,
  norm_shift = NULL, z_interval = 1, point_size = 2,
  plane_col = "black", plane_alpha = 0.1, border_col = "black",
  border_alpha = 1, border_lwd = 1, segment_col = NULL,
  segment_alpha = 0.3, segment_lwd = 1, new_device = FALSE,
  reset_view = TRUE, userMatrix = rotationMatrix(pi/2, 0, 1, 0))
}
\arguments{
\item{x}{x}
}
\description{
grimon

Grimon (Graphical interface to visualize multi-omics networks)
visualizes high-dimensional multi-layered data sets in three-dimensional
parallel coordinates. It enables users to intuitively and interactively
explore their data, helping their understanding of multiple inter-layer
connections embedded in high-dimensional complex data.
}
\examples{
data("grimon.example")
grimon(x, col = col, label = 1:6,
       optimize_coordinates = TRUE, maxiter = 1e3,
       segment_alpha = 0.5)
}
\author{
Masahiro Kanai
}
